Had to fix the AppSettingsStore get_reference().
By definition, this totally does the wrong thing and gets you in a
trap.  You must do new on the object.  For one, the get_reference
call is hard coded to the base class, so you cannot get YOUR derived
object returned.  Second, you cannot override it, because then you
cannot call the internal stuff that happens down inside this function.

So we ended up doing this all through the inititalize methods.