Fix for parameter defaults.
The parent must call an instance method to properly iterate through all
the levels of the class hierarchy.  When calling a class method, it only
gets the one it knows about, which is generally itself, and thus fails
to walk the hierarchy.  The simplest way is to use the instance method.

However, instance methods won't allow the defaults to be examined
externally. So for now, we have the ugly hack of having both, which
nearly read the same.

What would probably work is for the original parent to use the getClass
method and then do reflection to obtain the actual class method for the
final child derivation.  Then we could have just a chain of class
methods and avoid the instance method entirely.  Need to look at
doing this.